requirements.txt
# Django及REST Framework
django>=4.2.0
djangorestframework>=3.14.0

# JWT认证
djangorestframework-simplejwt>=5.2.0

# 密码加密
pycryptodome>=3.15.0

# 环境变量管理
python-dotenv>=1.0.0

# Redis缓存
redis>=4.5.0